requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.7.4
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
motor==3.5.0
multidict==6.0.5
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==4.8.0
python-dotenv==1.0.1
six==1.16.0
typing_extensions==4.12.2
yarl==1.9.4
